OBPIH-6500 Move create invoice page component to invoice/create
